refactor: Remove old TLS 1.2 structs
